@@ -1,7 +1,7 @@
 {
   "name": "converjon",
   "description": "An advanced image conversion server and command line tool.",
-  "version": "2.7.1",
+  "version": "2.7.9",
   "license": "MIT",
   "author": "Leon Weidauer",
   "maintainers": [
@@ -9,7 +9,8 @@
   ],
   "contributors": [
     "Leon Weidauer <leon@lnwdr.de> (http://lnwdr.de)",
-    "Steffen Gransow <converjon@mivesto.de> (https://github.com/graste)"
+    "Steffen Gransow <converjon@mivesto.de> (https://github.com/graste)",
+    "https://github.com/berlinonline/converjon/graphs/contributors"
   ],
   "homepage": "https://github.com/berlinonline/converjon",
   "bugs": "https://github.com/berlinonline/converjon/issues",
@@ -32,23 +33,25 @@
   ],
   "dependencies": {
     "calmcard": "^0.1.1",
-    "cjson": "^0.3.0",
+    "cjson": "^0.3.3",
     "cli-color": "^0.3.3",
-    "clone": "^1.0.0",
-    "handlebars": "^3.0.0",
-    "js-yaml": "^3.0.2",
-    "minimist": "^1.1.2",
-    "mkdirp": "^0.5.0",
+    "clone": "^1.0.2",
+    "handlebars": "^4.0.11",
+    "js-yaml": "^3.8.1",
+    "minimist": "^1.2.0",
+    "mkdirp": "^0.5.1",
     "moniker": "^0.1.2",
-    "nodeunit": "^0.9.0",
-    "request": "^2.60.0",
-    "rsvp": "^3.0.20",
-    "send": "^0.9.3"
+    "request": "^2.79.0",
+    "rsvp": "^3.3.3",
+    "send": "^0.16.2"
   },
-  "devDependencies": {},
+  "devDependencies": {
+    "nodeunit": "^0.11.3",
+    "sshpk": "^1.13.2"
+  },
   "optionalDependencies": {},
   "engines": {
-    "node": ">=0.10.7 <0.13.0"
+    "node": ">=4.0.0 <12.1.0"
   },
   "scripts": {
     "test": "nodeunit ./test/*"
