@@ -116,8 +116,7 @@ function cache(item, options) {
         source_path = item.source;
     }
 
-    var target_path = pathutils.join([dir_path, pathutils.getOptionsPath(options)]);
-
+    var target_path = pathutils.join([dir_path, 'targets', pathutils.getOptionsPath(options)]);
 
     /*
      * lock the files in sequence, not in parallel
