@@ -68,14 +68,12 @@ var value_normalizers = {
 function normalize_values(cfg) {
     var i;
     var normalizers;
-    for (i in cfg) {
-        if (cfg.hasOwnProperty(i)) {
-            normalizers = value_normalizers[Object.prototype.toString.call(cfg[i])] || [];
-            cfg[i] = util.apply_functions(cfg[i], normalizers);
-        }
-    }
 
-    return cfg;
+    return Object.keys(cfg).reduce(function(carry, i) {
+        normalizers = value_normalizers[Object.prototype.toString.call(cfg[i])] || [];
+        carry[i] = util.apply_functions(cfg[i], normalizers);
+        return carry;
+    }, {});
 }
 
 /**
