@@ -7,6 +7,7 @@
 var path_sep = require("path").sep;
 
 var util = require("./util");
+var crypto = require("crypto");
 
 function join(parts) {
     return parts.join(path_sep);
@@ -23,40 +24,24 @@ function replaceAll(string, omit, place, prevstring) {
 exports.join = join;
 
 exports.getHashPath = function(input) {
-    var hash = replaceAll((new Buffer(input)).toString('base64'), "=", "");
-    var part;
-
-    var path_parts = [];
-
-    /*
-     * split the has string into chunks of max length 8
-     */
-    do {
-        part = hash.substr(0,8);
-        if (part !== "") {
-            path_parts.push(part);
-        }
-        hash = hash.substr(8);
-    } while (part !== "");
-
-    return join(path_parts);
+    var hash = crypto.createHash('sha256').update(Buffer.from(input)).digest("hex");
+    return hash.substr(0, 2)+path_sep+hash;
 };
 
 exports.getOptionsPath = function(options) {
-    var i, type;
-
     var path_parts = [];
+    var sortedProperties = util.sortProperties(options);
 
-    for (i in util.sortProperties(options)) {
-        type = typeof options[i];
-        if (type === "string" || type === "number") {
-            path_parts.push(i + "_" + options[i]);
-        }
-    }
+    path_parts = Object.keys(sortedProperties).filter(function(i) {
+        var type = typeof options[i];
+        return (type === "string" || type === "number");
+    }).map(function(i) {
+        return i + "_" + options[i];
+    });
 
     if (path_parts.length === 0) {
-        path_parts.push("__no_options");
+        return "__no_options";
     }
 
-    return join(path_parts);
+    return crypto.createHash('sha256').update(Buffer.from(join(path_parts))).digest("hex");
 };
