@@ -21,10 +21,10 @@ function set_default_options(item) {
 function extract_preset_data(item) {
     var options = item.options;
 
-    if (options.hasOwnProperty('preset')) {
+    if (typeof options.preset !== "undefined") {
 
-        if (!item.conf.presets.hasOwnProperty(options.preset)) {
-             throw new PresetError(options.preset);
+        if (typeof item.conf.presets[options.preset] === "undefined") {
+            throw new PresetError(options.preset);
         } else {
              item.options = merge(options, item.conf.presets[options.preset]);
         }
@@ -73,11 +73,9 @@ function create_aoi(w, h, aoi, defaults) {
         h: temp[3]
     };
 
-    for (var i in defaults) {
-        if (defaults.hasOwnProperty(i)) {
-            aoi[i] = defaults[i];
-        }
-    }
+    Object.keys(defaults).forEach(function(i) {
+        aoi[i] = defaults[i];
+    })
 
     return aoi;
 }
@@ -182,19 +180,19 @@ function check_constraints(item) {
     var i, c, o;
 
     for (i in  item.options) {
-        if (options.hasOwnProperty(i) && constraints.hasOwnProperty(i)) {
+        if (typeof options[i] !== "undefined" && typeof constraints[i] !== "undefined") {
             c = constraints[i];
             o = options[i];
 
             if (typeof c.type !== "undefined") {
                 if (c.type === "integer") {
-                    if (o !== parseInt(o, 10).toString()) {
+                    if (o.toString() !== parseInt(o, 10).toString()) {
                         throw new ConstraintError(i, o, c);
                     }
                 }
 
                 if (c.type === "float") {
-                    if (o !== parseFloat(o, 10).toString()) {
+                    if (o.toString() !== parseFloat(o, 10).toString()) {
                         throw new ConstraintError(i, o, c);
                     }
                 }
@@ -218,17 +216,17 @@ function check_constraints(item) {
 function check_padding_color(item) {
     var options = item.options;
 
-    if (options.hasOwnProperty("padding_color")) {
-      if (!/^#([a-fA-F0-9]{6}|[a-fA-F0-9]{3})$/.test(options.padding_color)) {
-        throw new ColorFormatError(options.padding_color);
-      }
+    if (typeof options.padding_color !== "undefined") {
+        if (!/^#([a-fA-F0-9]{6}|[a-fA-F0-9]{3})$/.test(options.padding_color)) {
+            throw new ColorFormatError(options.padding_color);
+        }
     }
 }
 
 function create_target_file(item) {
     set_default_options(item);
     var promise = new Promise(function(resolve, reject) {
-        logging.debug(item.id, "Applying cropping and constriants");
+        logging.debug(item.id, "Applying cropping and constraints");
         try {
             extract_preset_data(item);
             add_cropping(item);
