@@ -55,13 +55,12 @@ function add_original_headers(res, item) {
     var headers = item.meta_data.headers;
     var i;
 
-    for(i in headers) {
-        if (headers.hasOwnProperty(i)) {
-            if (original_header_whitelist.indexOf(i.toLowerCase()) >= 0) {
-                res.setHeader(i, headers[i]);
-            }
+    Object.keys(headers).reduce(function(carry, current) {
+        if (original_header_whitelist.indexOf(current.toLowerCase()) >= 0) {
+            res.setHeader(current, headers[current]);
         }
-    }
+        return res;
+    }, res)
 }
 
 function add_default_headers(res, item) {
