@@ -25,7 +25,7 @@ function source_http_error(err, code) {
     return err;
 }
 
-function analyze_result(item, headers) {
+function analyze_result(item) {
     var locks = item.locks;
     var id = item.id;
     var conf = item.conf;
@@ -183,7 +183,6 @@ function copy_source(item) {
     });
 }
 
-
 var get_source = function(item) {
     var source = item.source;
 
@@ -194,7 +193,7 @@ var get_source = function(item) {
     }
 
     var promise = fsutils.mkdirForFile(item.locks.meta.key).
-    then(function(dir_path) {
+                          then(function(dir_path) {
         return fsutils.mkdirForFile(item.locks.source.key);
     }).
     then(function(dir_path) {
@@ -202,14 +201,13 @@ var get_source = function(item) {
             case "url": return download_source(item);
             case "file":
                 item.source_headers = item.conf.headers || {};
-                item.source_headers.date = (new Date()).toUTCString();
+                item.source_headers.date = (new Date()).toUTCString(); //stat.mtime.toUTCString();
                 if (item.conf.cache.copy_source_file) {
                     return copy_source(item);
                 } else {
                     logging.debug(item.id, "Using original local file as source:", item.source);
                     return;
                 }
-                break;
             default: throw new Error("Invalid source type. Only 'url' and 'file' are allowed.");
         }
     }).
