@@ -10,20 +10,15 @@
  * @returns {object} the sorted version
  */
 exports.sortProperties = function (obj) {
-    var i;
-    var keys = [];
-    var sorted = {};
-    for (i in obj) {
-        if (obj.hasOwnProperty(i)) {
-            keys.push(i);
-        }
-    }
+    var sorted;
+    var keys = Object.keys(obj);
 
     keys.sort();
 
-    for (i = 0; i < keys.length; i++) {
-        sorted[keys[i]] = obj[keys[i]];
-    }
+    sorted = keys.reduce(function(carry, current) {
+        carry[current] = obj[current]
+        return carry;
+    }, {})
 
     return sorted;
 };
